Improved performance using PIL

+++ US_Aerial/Autoprep/GetAerialreworkPNG.py
@@ -149,6 +149,7 @@ if __name__ == '__main__':
             imagesource = image['IMAGE_SOURCE']
             imagecollection = image['IMAGE_COLLECTION_TYPE']
             originalpath = image['ORIGINAL_IMAGE_PATH']
+            imageuploadpath = originalpath
             if imagecollection == 'DOQQ':
                 arcpy.AddWarning('Cannot convert DOQQ image '+originalpath)
             else:
@@ -157,11 +158,16 @@ if __name__ == '__main__':
                     """Two copies are performed, 1 to convert the raster into a PNG for the application.
                     the other to only copy the image without spatial information to the job folder"""
                     #arcpy.CopyRaster_management(originalpath,os.path.join(scratch,job_image_name),colormap_to_RGB='ColormapToRGB',pixel_type='8_BIT_UNSIGNED',format='PNG',transform='NONE')
-                    arcpy.env.compression = "JPEG 50"
-                    arcpy.CopyRaster_management(originalpath,os.path.join(scratch,job_image_name),colormap_to_RGB='ColormapToRGB',pixel_type='8_BIT_UNSIGNED',format='JPEG',transform='NONE')
-                    if os.path.exists(os.path.join(job_folder,'gc',job_image_name)):
-                        os.remove(os.path.join(job_folder,'gc',job_image_name))
-                    shutil.copy(os.path.join(scratch,job_image_name),os.path.join(job_folder,'gc',job_image_name))
+                    if imageuploadpath[-3:] in ['jpg','png','tif']
+                        im = Image.open(imageuploadpath)
+                        subject_im = im.convert("RGB")
+                        subject_im.save(os.path.join(job_folder,'gc',job_image_name))
+                    else:
+                        arcpy.env.compression = "JPEG 50"
+                        arcpy.CopyRaster_management(originalpath,os.path.join(scratch,job_image_name),colormap_to_RGB='ColormapToRGB',pixel_type='8_BIT_UNSIGNED',format='JPEG',transform='NONE')
+                        if os.path.exists(os.path.join(job_folder,'gc',job_image_name)):
+                            os.remove(os.path.join(job_folder,'gc',job_image_name))
+                        shutil.copy(os.path.join(scratch,job_image_name),os.path.join(job_folder,'gc',job_image_name))
                 elif not os.path.exists(originalpath):
                     arcpy.AddError('cannot find image in inventory to convert, PLEASE CHECK PATH: '+originalpath)
     except Exception as e:
